+++ src/main/java/dev/spagurder/bribery/Bribery.java
@@ -1,7 +1,10 @@
 package dev.spagurder.bribery;
 
 import dev.spagurder.bribery.config.Config;
+import dev.spagurder.bribery.core.BribableEntityRegistry;
 import eu.midnightdust.lib.config.MidnightConfig;
+import net.minecraft.world.entity.animal.IronGolem;
+import net.minecraft.world.entity.npc.Villager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -14,6 +17,15 @@ public class Bribery {
         LOGGER.info("Initializing Bribery");
         MidnightConfig.init(MOD_ID, Config.class);
         Config.refreshCurrencyItems();
+        loadEntities();
+    }
+
+    private static void loadEntities() {
+        BribableEntityRegistry.BRIBABLE_ENTITIES.add(Villager.class);
+        BribableEntityRegistry.BRIBABLE_ENTITIES.add(IronGolem.class);
+        BribableEntityRegistry.tryLoad("tallestegg.guardvillagers.common.entities.Guard");
+        BribableEntityRegistry.tryLoad("tallestegg.guardvillagers.entities.Guard");
+        BribableEntityRegistry.tryLoad("dev.sterner.guardvillagers.common.entity.GuardEntity");
     }
 
 }