+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 - Move transient config fields to separate class (fix GSON errors)
 - Add Thief Compat
+- Add Guard Villagers Compat
+- Neoforge Bugfix